Flask/requirements.txt
click==7.1.1
Flask==1.1.1
Flask-WTF==0.14.3
gitdb==4.0.4
GitPython==3.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
paho-mqtt==1.5.0
pkg-resources==0.0.0
PyMySQL==0.9.3
smmap==3.0.2
Werkzeug==1.0.0
WTForms==2.2.1
